doing truncate

+++ src/main/java/GrupoA/FuseSupport/CephishFuse.java
@@ -112,6 +112,12 @@ public class CephishFuse extends FuseStubFS {
 
     @Override
     public int unlink(String path) {
+        NodeAttributes attr = restClient.getAttribute(path);
+        if (attr == null)
+            return -ErrorCodes.ENOENT();
+        if(attr._FileType.equals(FileType.DIR))
+            return -ErrorCodes.EISDIR();
+
         return restClient.removeFile(path);
     }
 
@@ -168,6 +174,17 @@ public class CephishFuse extends FuseStubFS {
         System.out.println("----------------------------------------------------------");
 
         return writeResult;
+    }
+
+    @Override
+    public int truncate(String path, @off_t long size) {
+        NodeAttributes attr = restClient.getAttribute(path);
+        if (attr == null)
+            return -ErrorCodes.ENOENT();
+        if (attr._FileType.equals(FileType.DIR))
+            return -ErrorCodes.EISDIR();
+
+        return restClient.truncate(path, size);
     }
 /*
     @Override
